+++ src/components/general-widgets/CallMeBack/component.js
@@ -27,25 +27,27 @@ import { ICON_NAMES } from 'elements/Icon';
  * @returns {Object}
  */
 export const Component = ({
-  datePlaceholderText,
+  dateInputPlaceholder,
   emailInputLabel,
   headingText,
   nameInputLabel,
-  noteTextareaLabel,
+  notesInputLabel,
   onSubmit,
   phoneInputLabel,
-  propertyDropdownLabel,
+  propertyInputLabel,
   propertyOptions,
   submitButtonText,
-  timeDropdownLabel,
+  timeInputLabel,
   timeOptions,
-  timezoneDropdownLabel,
+  timeZoneInputLabel,
   timeZoneOptions,
+  validation,
 }) => (
   <Form
     headingText={headingText}
     onSubmit={onSubmit}
     submitButtonText={submitButtonText}
+    validation={validation}
   >
     <InputGroup>
       <TextInput label={nameInputLabel} name="name" />
@@ -53,65 +55,66 @@ export const Component = ({
     </InputGroup>
     <TextInput label={emailInputLabel} name="email" />
     <InputGroup>
-      <SingleDatePicker name="date" placeholderText={datePlaceholderText} />
+      <SingleDatePicker name="date" placeholderText={dateInputPlaceholder} />
       <Dropdown
         icon={ICON_NAMES.CLOCK}
-        label={timeDropdownLabel}
+        label={timeInputLabel}
         name="time"
         options={timeOptions}
       />
     </InputGroup>
     <InputGroup>
       <Dropdown
-        label={timezoneDropdownLabel}
+        label={timeZoneInputLabel}
         name="timeZone"
         options={timeZoneOptions}
       />
       <Dropdown
-        label={propertyDropdownLabel}
+        label={propertyInputLabel}
         name="property"
         options={propertyOptions}
       />
     </InputGroup>
-    <TextArea label={noteTextareaLabel} name="notes" />
+    <TextArea label={notesInputLabel} name="notes" />
   </Form>
 );
 
 Component.displayName = 'CallMeBack';
 
 Component.defaultProps = {
-  datePlaceholderText: DATE,
+  dateInputPlaceholder: DATE,
   emailInputLabel: EMAIL,
   headingText: CALL_ME_BACK,
   nameInputLabel: NAME,
-  noteTextareaLabel: NOTES,
+  notesInputLabel: NOTES,
   onSubmit: Function.prototype,
   phoneInputLabel: PHONE,
-  propertyDropdownLabel: PROPERTY,
+  propertyInputLabel: PROPERTY,
   submitButtonText: SEND,
-  timeDropdownLabel: TIME,
-  timezoneDropdownLabel: TIME_ZONE,
+  timeInputLabel: TIME,
+  timeZoneInputLabel: TIME_ZONE,
+  validation: {},
 };
 
 Component.propTypes = {
-  /** The placeholder text for the date input */
-  datePlaceholderText: PropTypes.string,
-  /** The label for the email input */
+  /** The placeholder for the date input. */
+  dateInputPlaceholder: PropTypes.string,
+  /** The label for the email input. */
   emailInputLabel: PropTypes.string,
-  /** The text for the heading displayed above the call me back form */
+  /** The text to display as a heading at the top of the widget. */
   headingText: PropTypes.string,
-  /** The label for the name input */
+  /** The label for the name input. */
   nameInputLabel: PropTypes.string,
-  /** The label for the note text area */
-  noteTextareaLabel: PropTypes.string,
+  /** The label for the notes input. */
+  notesInputLabel: PropTypes.string,
   /** The function to call when the form is submitted
    *  @param {Object} values - The values of the inputs in the form.
    */
   onSubmit: PropTypes.func,
   /** The label for the phone input */
   phoneInputLabel: PropTypes.string,
-  /** The label for the property dropdown */
-  propertyDropdownLabel: PropTypes.string,
+  /** The label for the property input. */
+  propertyInputLabel: PropTypes.string,
   /** The options which the user can select for the property field. */
   propertyOptions: PropTypes.arrayOf(
     PropTypes.shape({
@@ -127,8 +130,8 @@ Component.propTypes = {
   ).isRequired,
   /** The form submit button text */
   submitButtonText: PropTypes.string,
-  /** The label for the time dropdown */
-  timeDropdownLabel: PropTypes.string,
+  /** The label for the time input. */
+  timeInputLabel: PropTypes.string,
   /** The options which the user can select for the time field. */
   timeOptions: PropTypes.arrayOf(
     PropTypes.shape({
@@ -142,6 +145,8 @@ Component.propTypes = {
       ]),
     })
   ).isRequired,
+  /** The label for the time zone input. */
+  timeZoneInputLabel: PropTypes.string,
   /** The options which the user can select for the time zone field. */
   timeZoneOptions: PropTypes.arrayOf(
     PropTypes.shape({
@@ -155,6 +160,12 @@ Component.propTypes = {
       ]),
     })
   ).isRequired,
-  /** The label for the time zone dropdown */
-  timezoneDropdownLabel: PropTypes.string,
+  /** Settings for validating inputs. Each value should match [the shape documented in `Form`](https://lodgify.github.io/lodgify-ui/#/Collections/Form) */
+  validation: PropTypes.shape({
+    date: PropTypes.object,
+    email: PropTypes.object,
+    name: PropTypes.object,
+    notes: PropTypes.object,
+    phone: PropTypes.object,
+  }),
 };