+++ src/components/general-widgets/Contact/component.spec.js
@@ -54,6 +54,7 @@ describe('<Contact />', () => {
         headingText: CONTACT,
         onSubmit: Function.prototype,
         submitButtonText: SEND,
+        validation: expect.any(Object),
       });
     });