+++ src/components/general-widgets/OwnerLogin/utils/getForgotPasswordFormMarkup.js
@@ -1,32 +1,33 @@
 import React from 'react';
 
-import { FORGOT_PASSWORD, SEND_RESET, EMAIL } from 'utils/default-strings';
 import { Form } from 'collections/Form';
 import { Modal } from 'elements/Modal';
 import { TextInput } from 'inputs/TextInput';
 
 /**
  * @param {Function}  onForgotPasswordSubmit
- * @param {String}    forgotPasswordButtonText
- * @param {String}    forgotPasswordEmailLabel
+ * @param {String}    forgotPasswordSubmitButtonText
+ * @param {String}    forgotPasswordEmailInputLabel
  * @param {String}    forgotPasswordHeadingText
- * @param {String}    forgotPasswordModelTriggerText
+ * @param {String}    forgotPasswordModalTriggerText
  * @return {Object}
  */
 export const getForgotPasswordFormMarkup = (
   onForgotPasswordSubmit,
-  forgotPasswordButtonText = SEND_RESET,
-  forgotPasswordEmailLabel = EMAIL,
-  forgotPasswordHeadingText = FORGOT_PASSWORD,
-  forgotPasswordModelTriggerText = FORGOT_PASSWORD
+  forgotPasswordSubmitButtonText,
+  forgotPasswordEmailInputLabel,
+  forgotPasswordHeadingText,
+  forgotPasswordModalTriggerText,
+  forgotPasswordValidation
 ) => (
-  <Modal trigger={<span>{forgotPasswordModelTriggerText}</span>}>
+  <Modal trigger={<span>{forgotPasswordModalTriggerText}</span>}>
     <Form
       headingText={forgotPasswordHeadingText}
       onSubmit={onForgotPasswordSubmit}
-      submitButtonText={forgotPasswordButtonText}
+      submitButtonText={forgotPasswordSubmitButtonText}
+      validation={forgotPasswordValidation}
     >
-      <TextInput label={forgotPasswordEmailLabel} name="email" />
+      <TextInput label={forgotPasswordEmailInputLabel} name="email" />
     </Form>
   </Modal>
 );