+++ src/components/property-page-widgets/HostProfile/utils/getStringWithColonSuffix.spec.js
@@ -1,0 +1,10 @@
+import { getStringWithColonSuffix } from './getStringWithColonSuffix';
+
+describe('`getStringWithColonSuffix`', () => {
+  it('should return the right string', () => {
+    const string = '🏰';
+    const actual = getStringWithColonSuffix(string);
+
+    expect(actual).toBe(`${string}: `);
+  });
+});