+++ src/components/property-page-widgets/Rates/utils/buildPricePerExtraGuestString.js
@@ -1,6 +1,7 @@
 /**
  * @param {String} cost
+ * @param {String} costPerExtraGuestLabel
  * @return {String}
  */
-export const buildPricePerExtraGuestString = cost =>
-  `${cost} / Price Per extra per.`;
+export const buildPricePerExtraGuestString = (cost, costPerExtraGuestLabel) =>
+  `${cost} / ${costPerExtraGuestLabel}`;