+++ src/components/property-page-widgets/Rates/utils/getStringWithColonSuffix.js
@@ -1,0 +1,5 @@
+/**
+ * @param  {String} string
+ * @return {String}
+ */
+export const getStringWithColonSuffix = string => `${string}: `;