+++ src/components/property-page-widgets/Rules/utils/getLabelAndValueString.js
@@ -1,0 +1,6 @@
+/**
+ * @param  {String} label
+ * @param  {String} value
+ * @return {String}
+ */
+export const getLabelAndValueString = (label, value) => `${label}: ${value}`;